imported fonts and images basic css
