+++ client/src/containers/MainContainer.jsx
@@ -82,7 +82,7 @@ export default function MainContainer(props) {
         <JobForm currentUser={currentUser} jobs={jobs} handleCreateJob={handleCreateJob} />
       </Route>
       <Route exact path='/users/:user_id/jobs'>
-        <UserJobs currentUser={currentUser} jobs={jobs} setJobs={setJobs}handleEditJob={handleEditJob} handleDeleteJob={handleDeleteJob}/>
+        <UserJobs currentUser={currentUser} jobs={jobs} handleEditJob={handleEditJob} handleDeleteJob={handleDeleteJob}/>
       </Route>
       <Route exact path='/jobs/:id'>
         <Confirmation currentUser={currentUser} jobs={jobs}/>
